+++ Napoj.hpp
@@ -11,7 +11,7 @@ class Napoj : public Jidlo
     int objem;
 
 public:
-    Napoj(int id, string nazev, int cena, int trida, string ingredience, int objem)
+    Napoj(int id, string nazev, double cena, int trida, string ingredience, int objem)
         : Jidlo(id, nazev, cena, trida, ingredience)
         , objem(objem){}; // inline definice - prazdne telo metody, zakoncene strednikem
 
@@ -19,7 +19,7 @@ public:
 
     int getId();
     string getNazev();
-    int getPrice();
+    double getPrice();
     int getTrida();
     string getIngredience();
     int getObjem();
